changed code after PR
