+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:sport_matcher/features/authentication/widget/authentication_screen.dart';
+import 'package:sport_matcher/ui/authentication/widgets/authentication_screen.dart';
 
 void main() {
   runApp(MaterialApp(
