+++ lib/ui/authentication/assets/logo.png
